初期化のメソッド分離。BASEはfactoryImpl経由せずにテストが出来る筈！
BASEクラス？はStrategyパターンに変更したので、Hook必要なハズ。テストドライバー側をもう少し見直したい。
